ml_service/requirements.txt
fastapi
uvicorn
Pillow
tensorflow
numpy
python-multipart
python-multipart

